requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
pymupdf>=1.19.0
pytesseract>=0.3.8
pillow>=8.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
